third_party/WebKit/LayoutTests/editing/selection/addRange-merging-expected.txt

CONSOLE ERROR: line 132: Discontiguous selection is not supported.
CONSOLE ERROR: line 132: Discontiguous selection is not supported.
CONSOLE ERROR: line 132: Discontiguous selection is not supported.
CONSOLE ERROR: line 132: Discontiguous selection is not supported.
CONSOLE ERROR: line 132: Discontiguous selection is not supported.
CONSOLE ERROR: line 132: Discontiguous selection is not supported.
CONSOLE ERROR: line 132: Discontiguous selection is not supported.
CONSOLE ERROR: line 132: Discontiguous selection is not supported.
CONSOLE ERROR: line 132: Discontiguous selection is not supported.
CONSOLE ERROR: line 132: Discontiguous selection is not supported.
CONSOLE ERROR: line 139: Discontiguous selection is not supported.
CONSOLE ERROR: line 139: Discontiguous selection is not supported.
CONSOLE ERROR: line 139: Discontiguous selection is not supported.
CONSOLE ERROR: line 139: Discontiguous selection is not supported.
CONSOLE ERROR: line 139: Discontiguous selection is not supported.
CONSOLE ERROR: line 139: Discontiguous selection is not supported.
CONSOLE ERROR: line 139: Discontiguous selection is not supported.
CONSOLE ERROR: line 139: Discontiguous selection is not supported.
CONSOLE ERROR: line 139: Discontiguous selection is not supported.
CONSOLE ERROR: line 139: Discontiguous selection is not supported.
CONSOLE ERROR: line 146: Discontiguous selection is not supported.
CONSOLE ERROR: line 146: Discontiguous selection is not supported.
CONSOLE ERROR: line 146: Discontiguous selection is not supported.
CONSOLE ERROR: line 146: Discontiguous selection is not supported.
CONSOLE ERROR: line 146: Discontiguous selection is not supported.
CONSOLE ERROR: line 146: Discontiguous selection is not supported.
CONSOLE ERROR: line 146: Discontiguous selection is not supported.
CONSOLE ERROR: line 146: Discontiguous selection is not supported.
CONSOLE ERROR: line 146: Discontiguous selection is not supported.
CONSOLE ERROR: line 146: Discontiguous selection is not supported.
CONSOLE ERROR: line 153: Discontiguous selection is not supported.
CONSOLE ERROR: line 153: Discontiguous selection is not supported.
CONSOLE ERROR: line 153: Discontiguous selection is not supported.
CONSOLE ERROR: line 153: Discontiguous selection is not supported.
CONSOLE ERROR: line 153: Discontiguous selection is not supported.
CONSOLE ERROR: line 153: Discontiguous selection is not supported.
CONSOLE ERROR: line 153: Discontiguous selection is not supported.
CONSOLE ERROR: line 153: Discontiguous selection is not supported.
CONSOLE ERROR: line 153: Discontiguous selection is not supported.
CONSOLE ERROR: line 153: Discontiguous selection is not supported.
Selection.addRange() should properly merge intersecting (and don't merge discrete) ranges.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Running: testExpandLeftToRight (initializePositionsFunction = initializeTextPositions, containerIsEditable = false)
PASS Selection was: start = ([Text: 12345], 1), end = ([Text: 12345], 4)
Running: testExpandLeftToRight (initializePositionsFunction = initializeTextPositions, containerIsEditable = true)
PASS Selection was: start = ([Text: 12345], 1), end = ([Text: 12345], 4)
Running: testExpandLeftToRight (initializePositionsFunction = initializeOuterElementPositions, containerIsEditable = false)
PASS Selection was: start = ([<DIV>: #container], 1), end = ([<DIV>: #container], 4)
Running: testExpandLeftToRight (initializePositionsFunction = initializeOuterElementPositions, containerIsEditable = true)
PASS Selection was: start = ([<DIV>: #container], 1), end = ([<DIV>: #container], 4)
Running: testExpandLeftToRight (initializePositionsFunction = initializeInnerElementPositions, containerIsEditable = false)
PASS Selection was: start = ([<SPAN>: #b], 0), end = ([<SPAN>: #e], 0)
Running: testExpandLeftToRight (initializePositionsFunction = initializeInnerElementPositions, containerIsEditable = true)
PASS Selection was: start = ([<SPAN>: #b], 0), end = ([<SPAN>: #e], 0)
Running: testExpandLeftToRight (initializePositionsFunction = initializeVisiblyEquivalentPositionsBeforeNodes, containerIsEditable = false)
PASS Selection was: start = ([<DIV>: #container], 0), end = ([<SPAN>: #c], 0)
Running: testExpandLeftToRight (initializePositionsFunction = initializeVisiblyEquivalentPositionsBeforeNodes, containerIsEditable = true)
PASS Selection was: start = ([<DIV>: #container], 0), end = ([<SPAN>: #c], 0)
Running: testExpandLeftToRight (initializePositionsFunction = initializeVisiblyEquivalentPositionsAfterNodes, containerIsEditable = false)
PASS Selection was: start = ([<SPAN>: #c], 0), end = ([<DIV>: #container], 1)
Running: testExpandLeftToRight (initializePositionsFunction = initializeVisiblyEquivalentPositionsAfterNodes, containerIsEditable = true)
PASS Selection was: start = ([<SPAN>: #c], 0), end = ([<DIV>: #container], 1)
Running: testExpandRightToLeft (initializePositionsFunction = initializeTextPositions, containerIsEditable = false)
PASS Selection was: start = ([Text: 12345], 1), end = ([Text: 12345], 4)
Running: testExpandRightToLeft (initializePositionsFunction = initializeTextPositions, containerIsEditable = true)
PASS Selection was: start = ([Text: 12345], 1), end = ([Text: 12345], 4)
Running: testExpandRightToLeft (initializePositionsFunction = initializeOuterElementPositions, containerIsEditable = false)
PASS Selection was: start = ([<DIV>: #container], 1), end = ([<DIV>: #container], 4)
Running: testExpandRightToLeft (initializePositionsFunction = initializeOuterElementPositions, containerIsEditable = true)
PASS Selection was: start = ([<DIV>: #container], 1), end = ([<DIV>: #container], 4)
Running: testExpandRightToLeft (initializePositionsFunction = initializeInnerElementPositions, containerIsEditable = false)
PASS Selection was: start = ([<SPAN>: #b], 0), end = ([<SPAN>: #e], 0)
Running: testExpandRightToLeft (initializePositionsFunction = initializeInnerElementPositions, containerIsEditable = true)
PASS Selection was: start = ([<SPAN>: #b], 0), end = ([<SPAN>: #e], 0)
Running: testExpandRightToLeft (initializePositionsFunction = initializeVisiblyEquivalentPositionsBeforeNodes, containerIsEditable = false)
PASS Selection was: start = ([<DIV>: #container], 0), end = ([<SPAN>: #c], 0)
Running: testExpandRightToLeft (initializePositionsFunction = initializeVisiblyEquivalentPositionsBeforeNodes, containerIsEditable = true)
PASS Selection was: start = ([<DIV>: #container], 0), end = ([<SPAN>: #c], 0)
Running: testExpandRightToLeft (initializePositionsFunction = initializeVisiblyEquivalentPositionsAfterNodes, containerIsEditable = false)
PASS Selection was: start = ([<SPAN>: #c], 0), end = ([<DIV>: #container], 1)
Running: testExpandRightToLeft (initializePositionsFunction = initializeVisiblyEquivalentPositionsAfterNodes, containerIsEditable = true)
PASS Selection was: start = ([<SPAN>: #c], 0), end = ([<DIV>: #container], 1)
Running: testExpandLeftToRightAdjacent (initializePositionsFunction = initializeTextPositions, containerIsEditable = false)
PASS Selection was: start = ([Text: 12345], 2), end = ([Text: 12345], 4)
Running: testExpandLeftToRightAdjacent (initializePositionsFunction = initializeTextPositions, containerIsEditable = true)
PASS Selection was: start = ([Text: 12345], 2), end = ([Text: 12345], 4)
Running: testExpandLeftToRightAdjacent (initializePositionsFunction = initializeOuterElementPositions, containerIsEditable = false)
PASS Selection was: start = ([<DIV>: #container], 2), end = ([<DIV>: #container], 4)
Running: testExpandLeftToRightAdjacent (initializePositionsFunction = initializeOuterElementPositions, containerIsEditable = true)
PASS Selection was: start = ([<DIV>: #container], 2), end = ([<DIV>: #container], 4)
Running: testExpandLeftToRightAdjacent (initializePositionsFunction = initializeInnerElementPositions, containerIsEditable = false)
PASS Selection was: start = ([<SPAN>: #c], 0), end = ([<SPAN>: #e], 0)
Running: testExpandLeftToRightAdjacent (initializePositionsFunction = initializeInnerElementPositions, containerIsEditable = true)
PASS Selection was: start = ([<SPAN>: #c], 0), end = ([<SPAN>: #e], 0)
Running: testExpandLeftToRightAdjacent (initializePositionsFunction = initializeVisiblyEquivalentPositionsBeforeNodes, containerIsEditable = false)
PASS Selection was: start = ([<SPAN>: #a], 0), end = ([<SPAN>: #c], 0)
Running: testExpandLeftToRightAdjacent (initializePositionsFunction = initializeVisiblyEquivalentPositionsBeforeNodes, containerIsEditable = true)
PASS Selection was: start = ([<SPAN>: #a], 0), end = ([<SPAN>: #c], 0)
Running: testExpandLeftToRightAdjacent (initializePositionsFunction = initializeVisiblyEquivalentPositionsAfterNodes, containerIsEditable = false)
PASS Selection was: start = ([<SPAN>: #b], 1), end = ([<DIV>: #container], 1)
Running: testExpandLeftToRightAdjacent (initializePositionsFunction = initializeVisiblyEquivalentPositionsAfterNodes, containerIsEditable = true)
PASS Selection was: start = ([<SPAN>: #b], 1), end = ([<DIV>: #container], 1)
Running: testExpandRightToLeftAdjacent (initializePositionsFunction = initializeTextPositions, containerIsEditable = false)
PASS Selection was: start = ([Text: 12345], 1), end = ([Text: 12345], 3)
Running: testExpandRightToLeftAdjacent (initializePositionsFunction = initializeTextPositions, containerIsEditable = true)
PASS Selection was: start = ([Text: 12345], 1), end = ([Text: 12345], 3)
Running: testExpandRightToLeftAdjacent (initializePositionsFunction = initializeOuterElementPositions, containerIsEditable = false)
PASS Selection was: start = ([<DIV>: #container], 1), end = ([<DIV>: #container], 3)
Running: testExpandRightToLeftAdjacent (initializePositionsFunction = initializeOuterElementPositions, containerIsEditable = true)
PASS Selection was: start = ([<DIV>: #container], 1), end = ([<DIV>: #container], 3)
Running: testExpandRightToLeftAdjacent (initializePositionsFunction = initializeInnerElementPositions, containerIsEditable = false)
PASS Selection was: start = ([<SPAN>: #b], 0), end = ([<SPAN>: #d], 0)
Running: testExpandRightToLeftAdjacent (initializePositionsFunction = initializeInnerElementPositions, containerIsEditable = true)
PASS Selection was: start = ([<SPAN>: #b], 0), end = ([<SPAN>: #d], 0)
Running: testExpandRightToLeftAdjacent (initializePositionsFunction = initializeVisiblyEquivalentPositionsBeforeNodes, containerIsEditable = false)
PASS Selection was: start = ([<DIV>: #container], 0), end = ([<SPAN>: #b], 0)
Running: testExpandRightToLeftAdjacent (initializePositionsFunction = initializeVisiblyEquivalentPositionsBeforeNodes, containerIsEditable = true)
PASS Selection was: start = ([<DIV>: #container], 0), end = ([<SPAN>: #b], 0)
Running: testExpandRightToLeftAdjacent (initializePositionsFunction = initializeVisiblyEquivalentPositionsAfterNodes, containerIsEditable = false)
PASS Selection was: start = ([<SPAN>: #c], 0), end = ([<SPAN>: #a], 1)
Running: testExpandRightToLeftAdjacent (initializePositionsFunction = initializeVisiblyEquivalentPositionsAfterNodes, containerIsEditable = true)
PASS Selection was: start = ([<SPAN>: #c], 0), end = ([<SPAN>: #a], 1)
Running: testExpandBothEnds (initializePositionsFunction = initializeTextPositions, containerIsEditable = false)
PASS Selection was: start = ([Text: 12345], 1), end = ([Text: 12345], 4)
Running: testExpandBothEnds (initializePositionsFunction = initializeTextPositions, containerIsEditable = true)
PASS Selection was: start = ([Text: 12345], 1), end = ([Text: 12345], 4)
Running: testExpandBothEnds (initializePositionsFunction = initializeOuterElementPositions, containerIsEditable = false)
PASS Selection was: start = ([<DIV>: #container], 1), end = ([<DIV>: #container], 4)
Running: testExpandBothEnds (initializePositionsFunction = initializeOuterElementPositions, containerIsEditable = true)
PASS Selection was: start = ([<DIV>: #container], 1), end = ([<DIV>: #container], 4)
Running: testExpandBothEnds (initializePositionsFunction = initializeInnerElementPositions, containerIsEditable = false)
PASS Selection was: start = ([<SPAN>: #b], 0), end = ([<SPAN>: #e], 0)
Running: testExpandBothEnds (initializePositionsFunction = initializeInnerElementPositions, containerIsEditable = true)
PASS Selection was: start = ([<SPAN>: #b], 0), end = ([<SPAN>: #e], 0)
Running: testExpandBothEnds (initializePositionsFunction = initializeVisiblyEquivalentPositionsBeforeNodes, containerIsEditable = false)
PASS Selection was: start = ([<DIV>: #container], 0), end = ([<SPAN>: #c], 0)
Running: testExpandBothEnds (initializePositionsFunction = initializeVisiblyEquivalentPositionsBeforeNodes, containerIsEditable = true)
PASS Selection was: start = ([<DIV>: #container], 0), end = ([<SPAN>: #c], 0)
Running: testExpandBothEnds (initializePositionsFunction = initializeVisiblyEquivalentPositionsAfterNodes, containerIsEditable = false)
PASS Selection was: start = ([<SPAN>: #c], 0), end = ([<DIV>: #container], 1)
Running: testExpandBothEnds (initializePositionsFunction = initializeVisiblyEquivalentPositionsAfterNodes, containerIsEditable = true)
PASS Selection was: start = ([<SPAN>: #c], 0), end = ([<DIV>: #container], 1)
Running: testDontExpand (initializePositionsFunction = initializeTextPositions, containerIsEditable = false)
PASS Selection was: start = ([Text: 12345], 1), end = ([Text: 12345], 4)
Running: testDontExpand (initializePositionsFunction = initializeTextPositions, containerIsEditable = true)
PASS Selection was: start = ([Text: 12345], 1), end = ([Text: 12345], 4)
Running: testDontExpand (initializePositionsFunction = initializeOuterElementPositions, containerIsEditable = false)
PASS Selection was: start = ([<DIV>: #container], 1), end = ([<DIV>: #container], 4)
Running: testDontExpand (initializePositionsFunction = initializeOuterElementPositions, containerIsEditable = true)
PASS Selection was: start = ([<DIV>: #container], 1), end = ([<DIV>: #container], 4)
Running: testDontExpand (initializePositionsFunction = initializeInnerElementPositions, containerIsEditable = false)
PASS Selection was: start = ([<SPAN>: #b], 0), end = ([<SPAN>: #e], 0)
Running: testDontExpand (initializePositionsFunction = initializeInnerElementPositions, containerIsEditable = true)
PASS Selection was: start = ([<SPAN>: #b], 0), end = ([<SPAN>: #e], 0)
Running: testDontExpand (initializePositionsFunction = initializeVisiblyEquivalentPositionsBeforeNodes, containerIsEditable = false)
PASS Selection was: start = ([<DIV>: #container], 0), end = ([<SPAN>: #c], 0)
Running: testDontExpand (initializePositionsFunction = initializeVisiblyEquivalentPositionsBeforeNodes, containerIsEditable = true)
PASS Selection was: start = ([<DIV>: #container], 0), end = ([<SPAN>: #c], 0)
Running: testDontExpand (initializePositionsFunction = initializeVisiblyEquivalentPositionsAfterNodes, containerIsEditable = false)
PASS Selection was: start = ([<SPAN>: #c], 0), end = ([<DIV>: #container], 1)
Running: testDontExpand (initializePositionsFunction = initializeVisiblyEquivalentPositionsAfterNodes, containerIsEditable = true)
PASS Selection was: start = ([<SPAN>: #c], 0), end = ([<DIV>: #container], 1)
Running: testAddSameRange (initializePositionsFunction = initializeTextPositions, containerIsEditable = false)
PASS Selection was: start = ([Text: 12345], 2), end = ([Text: 12345], 3)
Running: testAddSameRange (initializePositionsFunction = initializeTextPositions, containerIsEditable = true)
PASS Selection was: start = ([Text: 12345], 2), end = ([Text: 12345], 3)
Running: testAddSameRange (initializePositionsFunction = initializeOuterElementPositions, containerIsEditable = false)
PASS Selection was: start = ([<DIV>: #container], 2), end = ([<DIV>: #container], 3)
Running: testAddSameRange (initializePositionsFunction = initializeOuterElementPositions, containerIsEditable = true)
PASS Selection was: start = ([<DIV>: #container], 2), end = ([<DIV>: #container], 3)
Running: testAddSameRange (initializePositionsFunction = initializeInnerElementPositions, containerIsEditable = false)
PASS Selection was: start = ([<SPAN>: #c], 0), end = ([<SPAN>: #d], 0)
Running: testAddSameRange (initializePositionsFunction = initializeInnerElementPositions, containerIsEditable = true)
PASS Selection was: start = ([<SPAN>: #c], 0), end = ([<SPAN>: #d], 0)
Running: testAddSameRange (initializePositionsFunction = initializeVisiblyEquivalentPositionsBeforeNodes, containerIsEditable = false)
PASS Selection was: start = ([<SPAN>: #a], 0), end = ([<SPAN>: #b], 0)
Running: testAddSameRange (initializePositionsFunction = initializeVisiblyEquivalentPositionsBeforeNodes, containerIsEditable = true)
PASS Selection was: start = ([<SPAN>: #a], 0), end = ([<SPAN>: #b], 0)
Running: testAddSameRange (initializePositionsFunction = initializeVisiblyEquivalentPositionsAfterNodes, containerIsEditable = false)
PASS Selection was: start = ([<SPAN>: #b], 1), end = ([<SPAN>: #a], 1)
Running: testAddSameRange (initializePositionsFunction = initializeVisiblyEquivalentPositionsAfterNodes, containerIsEditable = true)
PASS Selection was: start = ([<SPAN>: #b], 1), end = ([<SPAN>: #a], 1)
Running: testRejectDistantRangeAtRight (initializePositionsFunction = initializeTextPositions, containerIsEditable = false)
PASS Selection was: start = ([Text: 12345], 1), end = ([Text: 12345], 2)
Running: testRejectDistantRangeAtRight (initializePositionsFunction = initializeTextPositions, containerIsEditable = true)
PASS Selection was: start = ([Text: 12345], 1), end = ([Text: 12345], 2)
Running: testRejectDistantRangeAtRight (initializePositionsFunction = initializeOuterElementPositions, containerIsEditable = false)
PASS Selection was: start = ([<DIV>: #container], 1), end = ([<DIV>: #container], 2)
Running: testRejectDistantRangeAtRight (initializePositionsFunction = initializeOuterElementPositions, containerIsEditable = true)
PASS Selection was: start = ([<DIV>: #container], 1), end = ([<DIV>: #container], 2)
Running: testRejectDistantRangeAtRight (initializePositionsFunction = initializeInnerElementPositions, containerIsEditable = false)
PASS Selection was: start = ([<SPAN>: #b], 0), end = ([<SPAN>: #c], 0)
Running: testRejectDistantRangeAtRight (initializePositionsFunction = initializeInnerElementPositions, containerIsEditable = true)
PASS Selection was: start = ([<SPAN>: #b], 0), end = ([<SPAN>: #c], 0)
Running: testRejectDistantRangeAtRight (initializePositionsFunction = initializeVisiblyEquivalentPositionsBeforeNodes, containerIsEditable = false)
PASS Selection was: start = ([<DIV>: #container], 0), end = ([<SPAN>: #a], 0)
Running: testRejectDistantRangeAtRight (initializePositionsFunction = initializeVisiblyEquivalentPositionsBeforeNodes, containerIsEditable = true)
PASS Selection was: start = ([<DIV>: #container], 0), end = ([<SPAN>: #a], 0)
Running: testRejectDistantRangeAtRight (initializePositionsFunction = initializeVisiblyEquivalentPositionsAfterNodes, containerIsEditable = false)
PASS Selection was: start = ([<SPAN>: #c], 0), end = ([<SPAN>: #b], 1)
Running: testRejectDistantRangeAtRight (initializePositionsFunction = initializeVisiblyEquivalentPositionsAfterNodes, containerIsEditable = true)
PASS Selection was: start = ([<SPAN>: #c], 0), end = ([<SPAN>: #b], 1)
Running: testRejectDistantRangeAtLeft (initializePositionsFunction = initializeTextPositions, containerIsEditable = false)
PASS Selection was: start = ([Text: 12345], 3), end = ([Text: 12345], 4)
Running: testRejectDistantRangeAtLeft (initializePositionsFunction = initializeTextPositions, containerIsEditable = true)
PASS Selection was: start = ([Text: 12345], 3), end = ([Text: 12345], 4)
Running: testRejectDistantRangeAtLeft (initializePositionsFunction = initializeOuterElementPositions, containerIsEditable = false)
PASS Selection was: start = ([<DIV>: #container], 3), end = ([<DIV>: #container], 4)
Running: testRejectDistantRangeAtLeft (initializePositionsFunction = initializeOuterElementPositions, containerIsEditable = true)
PASS Selection was: start = ([<DIV>: #container], 3), end = ([<DIV>: #container], 4)
Running: testRejectDistantRangeAtLeft (initializePositionsFunction = initializeInnerElementPositions, containerIsEditable = false)
PASS Selection was: start = ([<SPAN>: #d], 0), end = ([<SPAN>: #e], 0)
Running: testRejectDistantRangeAtLeft (initializePositionsFunction = initializeInnerElementPositions, containerIsEditable = true)
PASS Selection was: start = ([<SPAN>: #d], 0), end = ([<SPAN>: #e], 0)
Running: testRejectDistantRangeAtLeft (initializePositionsFunction = initializeVisiblyEquivalentPositionsBeforeNodes, containerIsEditable = false)
PASS Selection was: start = ([<SPAN>: #b], 0), end = ([<SPAN>: #c], 0)
Running: testRejectDistantRangeAtLeft (initializePositionsFunction = initializeVisiblyEquivalentPositionsBeforeNodes, containerIsEditable = true)
PASS Selection was: start = ([<SPAN>: #b], 0), end = ([<SPAN>: #c], 0)
Running: testRejectDistantRangeAtLeft (initializePositionsFunction = initializeVisiblyEquivalentPositionsAfterNodes, containerIsEditable = false)
PASS Selection was: start = ([<SPAN>: #a], 1), end = ([<DIV>: #container], 1)
Running: testRejectDistantRangeAtLeft (initializePositionsFunction = initializeVisiblyEquivalentPositionsAfterNodes, containerIsEditable = true)
PASS Selection was: start = ([<SPAN>: #a], 1), end = ([<DIV>: #container], 1)
Running: testRejectDistantCollapsedRangeAtRight (initializePositionsFunction = initializeTextPositions, containerIsEditable = false)
PASS Selection was: start = ([Text: 12345], 1), end = ([Text: 12345], 2)
Running: testRejectDistantCollapsedRangeAtRight (initializePositionsFunction = initializeTextPositions, containerIsEditable = true)
PASS Selection was: start = ([Text: 12345], 1), end = ([Text: 12345], 2)
Running: testRejectDistantCollapsedRangeAtRight (initializePositionsFunction = initializeOuterElementPositions, containerIsEditable = false)
PASS Selection was: start = ([<DIV>: #container], 1), end = ([<DIV>: #container], 2)
Running: testRejectDistantCollapsedRangeAtRight (initializePositionsFunction = initializeOuterElementPositions, containerIsEditable = true)
PASS Selection was: start = ([<DIV>: #container], 1), end = ([<DIV>: #container], 2)
Running: testRejectDistantCollapsedRangeAtRight (initializePositionsFunction = initializeInnerElementPositions, containerIsEditable = false)
PASS Selection was: start = ([<SPAN>: #b], 0), end = ([<SPAN>: #c], 0)
Running: testRejectDistantCollapsedRangeAtRight (initializePositionsFunction = initializeInnerElementPositions, containerIsEditable = true)
PASS Selection was: start = ([<SPAN>: #b], 0), end = ([<SPAN>: #c], 0)
Running: testRejectDistantCollapsedRangeAtRight (initializePositionsFunction = initializeVisiblyEquivalentPositionsBeforeNodes, containerIsEditable = false)
PASS Selection was: start = ([<DIV>: #container], 0), end = ([<SPAN>: #a], 0)
Running: testRejectDistantCollapsedRangeAtRight (initializePositionsFunction = initializeVisiblyEquivalentPositionsBeforeNodes, containerIsEditable = true)
PASS Selection was: start = ([<DIV>: #container], 0), end = ([<SPAN>: #a], 0)
Running: testRejectDistantCollapsedRangeAtRight (initializePositionsFunction = initializeVisiblyEquivalentPositionsAfterNodes, containerIsEditable = false)
PASS Selection was: start = ([<SPAN>: #c], 0), end = ([<SPAN>: #b], 1)
Running: testRejectDistantCollapsedRangeAtRight (initializePositionsFunction = initializeVisiblyEquivalentPositionsAfterNodes, containerIsEditable = true)
PASS Selection was: start = ([<SPAN>: #c], 0), end = ([<SPAN>: #b], 1)
Running: testRejectDistantCollapsedRangeAtLeft (initializePositionsFunction = initializeTextPositions, containerIsEditable = false)
PASS Selection was: start = ([Text: 12345], 3), end = ([Text: 12345], 4)
Running: testRejectDistantCollapsedRangeAtLeft (initializePositionsFunction = initializeTextPositions, containerIsEditable = true)
PASS Selection was: start = ([Text: 12345], 3), end = ([Text: 12345], 4)
Running: testRejectDistantCollapsedRangeAtLeft (initializePositionsFunction = initializeOuterElementPositions, containerIsEditable = false)
PASS Selection was: start = ([<DIV>: #container], 3), end = ([<DIV>: #container], 4)
Running: testRejectDistantCollapsedRangeAtLeft (initializePositionsFunction = initializeOuterElementPositions, containerIsEditable = true)
PASS Selection was: start = ([<DIV>: #container], 3), end = ([<DIV>: #container], 4)
Running: testRejectDistantCollapsedRangeAtLeft (initializePositionsFunction = initializeInnerElementPositions, containerIsEditable = false)
PASS Selection was: start = ([<SPAN>: #d], 0), end = ([<SPAN>: #e], 0)
Running: testRejectDistantCollapsedRangeAtLeft (initializePositionsFunction = initializeInnerElementPositions, containerIsEditable = true)
PASS Selection was: start = ([<SPAN>: #d], 0), end = ([<SPAN>: #e], 0)
Running: testRejectDistantCollapsedRangeAtLeft (initializePositionsFunction = initializeVisiblyEquivalentPositionsBeforeNodes, containerIsEditable = false)
PASS Selection was: start = ([<SPAN>: #b], 0), end = ([<SPAN>: #c], 0)
Running: testRejectDistantCollapsedRangeAtLeft (initializePositionsFunction = initializeVisiblyEquivalentPositionsBeforeNodes, containerIsEditable = true)
PASS Selection was: start = ([<SPAN>: #b], 0), end = ([<SPAN>: #c], 0)
Running: testRejectDistantCollapsedRangeAtLeft (initializePositionsFunction = initializeVisiblyEquivalentPositionsAfterNodes, containerIsEditable = false)
PASS Selection was: start = ([<SPAN>: #a], 1), end = ([<DIV>: #container], 1)
Running: testRejectDistantCollapsedRangeAtLeft (initializePositionsFunction = initializeVisiblyEquivalentPositionsAfterNodes, containerIsEditable = true)
PASS Selection was: start = ([<SPAN>: #a], 1), end = ([<DIV>: #container], 1)
PASS successfullyParsed is true

TEST COMPLETE
